requirements.txt
streamlit>=1.27
tensorflow>=2.15,<2.17
pillow
numpy
